refactor: CustomAlertDialog 리팩터링
