+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/common/extension/FragmentExtension.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.view.common
+package com.bottari.presentation.common.extension
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
