+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/common/alart/DialogPresetType.kt
@@ -7,6 +7,11 @@ enum class DialogPresetType {
     ;
 
     fun applyTo(dialog: CustomAlertDialog) =
+        when (this) {
+            EXIT_WITHOUT_SAVE -> applyExitWithoutSave(dialog)
+        }
+
+    private fun applyExitWithoutSave(dialog: CustomAlertDialog) {
         with(dialog) {
             binding.tvDialogCustomTitle.setText(R.string.common_alert_dialog_title_text)
             binding.tvDialogCustomDescription.setText(R.string.common_alert_unsaved_dialog_message_text)
@@ -14,14 +19,13 @@ enum class DialogPresetType {
                 textRes = R.string.common_yes_btn_text,
                 textColorRes = R.color.white,
                 backgroundColorRes = R.color.primary,
-            ) { listener?.onClickPositive() }
-
+            )
             setNegativeButton(
                 textRes = R.string.common_no_btn_text,
                 textColorRes = R.color.gray_787878,
                 backgroundColorRes = R.color.white,
-            ) { listener?.onClickNegative() }
-
+            )
             setCloseButton()
         }
+    }
 }
